TF-245: Add the sequential combinator.
It's common to want to combine network layers sequentially. This commit adds
the >> operator that makes it trivial to feed the output of one layer to the
input of the next layer.

Note: this commit also adds in the skeleton for a parallel combinator (e.g.
one that could be easily used to form residual networks) but this triggers a
compiler crash. (TF-244)

Related bugs: TF-235