Add a trainer abstraction to provide a default training loop.
Because most supervised models have similar training loops, to help users get
started more easily, this commit adds a sketch for what a pre-built training
loop might look like. This will be used to help flesh out the API surface
based on real usage. The API should not be assumed to be frozen!

> Note: the implementation of the training loop should be completely re-written
> to be callback driven. It also should have reasonable defaults for outputing
> information to the console, and should have direct support for high
> performance input pipelines. These are considered out of scope for the
> initial check-in, in order to facilitate parallel development.