[Tests] Fix macOS tests by avoiding uses of 'copy' in 'TensorGroupTests'.
On macOS, `XCTestCase` is bridged from Apple Objective-C frameworks and has a method [`copy(with:)`](https://developer.apple.com/documentation/foundation/nscopying/1410311-copy) inherited from `NSCopying`. This caused an error:

> Use of 'copy' refers to class method 'copy(with:)' rather than global function 'copy(of:)' in module 'TensorFlowTests'.

This PR changes `copy(of:)` to an extension method `makeCopy()` for `TensorHandle` in the test file to avoid the funciton lookup issue.