Add `@_Freezable` property wrapper.
`@_Freezable` wraps differentiable values and provides toggleable
"trainability" via the `isFrozen` property and `freeze`/`unfreeze`
mutating methods.

When `isFrozen` is true, accesses to `value` have a derivative of zero.

The alternative name `@_Trainable` was considered, but `@_Trainable`
gives the impression that "only properties marked with `@_Trainable`
are trainable", which is not true: all properties are trainable (i.e.
their derivatives are propagated) by default.