Add default implementations of new requirements in TensorArrayProtocol (#290)
In the case of lazy tensor, these implementations trigger
materialization. Therefore, if the user needs to defer materialization in lazy
tensor, they will need to manually implement these methods without causing
materialization.

Note that this is a stop-gap solution while we work on the DerivedConformances
for TensorArrayProtocol and being added here to avoid any delay for v0.4
release.