Use control flow AD in layers and add leak checking utility.
Remove custom VJPs from all layers so that they go through control flow differentiation. This applies to `BatchNorm` and `Dropout`. `Dense` was already using control flow differentiation.

Introduce a tensor leak checking test utility `withTensorLeakChecking { ... }` and use it in tests. A `globalTensorCount` variable is added to the thread-local `Context` to keep track of the number of live tensors, and `TFETensorHandle`'s `init(_owning:)` and `deinit` increases and decreases `globalTensorCount`, respectively.

Resolves #389.