Rename Raw to _Raw, and provide a deprecated typealias. (#543)
Part of the work to [layer our APIs](https://docs.google.com/document/d/1HO_sMhZJHxlDqw4Pjz4qva2s5yzgwxel-82yoK8O6L4)
is to set up S4TF to have pluggable "backends" or runtimes. While exposing the
Raw ops is powerful, it is not portable. This change renames Raw to _Raw to
indicate it is an unstable / not-forever-guaranteed API.