CMake tests (#737)
* build: add some compatibility for CMake<3.17

Ensure that we can properly set `DT_RUNPATH` for libraries and
executables on older CMake releases.  This has been fixed upstream, but
rather than requiring an extremely new CMake, provide a compatibility
shim.

* Experimental: add Experimental module

This adds the experimental module and improves the install rules further
to support renamed modules and do less work in the case of an executable
target.

* build: add testing support

Enhance the CMake setup to be able to run the test suite.  This will
make it easier to enable tests for X10.