X10: adjustments to support Windows (#825)
Use `_aligned_malloc` and `_aligned_free` on Windows rather than the
POSIX `posix_memalign` or the C11 specific `aligned_alloc`.

Adjust the build rules to create target dependent libraries, fix the
linker options to be target dependent and add support for the import
library target needed for Windows.

The biggest change here is that users will now target `x10` rather than
the library name of `libx10.so`.  This will generate `libx10.so` on
Linux, `libx10.dylib` on macOS, and `x10.dll` on Windows.