x10: port for Darwin (#828)
The Xcode 11.3 compiler does objects to the use of the extension to
provide the overload for a std::operator<< in a different namespace.
Move the definition into the `std::` namespace.  Although this is
placing something into a restricted namespace, it is adding an overload
and is permissible.  This allows the streaming of the `grpc::Status`
type to the logging.

Adjust the allocation to use `posix_malloc` for the alignment rather
than the C11 `aligned_alloc` as the latter is unavailable in C++11 mode
and C11 needs to be explicitly opted into.  It should be available as
part of C++17 on macOS, however, tensorflow is still adhering C++11.