build: add `FindX10` (#829)
Add the ability to use an external build of X10.  This attempts to
search the swift toolchain in addition to the standard search order.
The user can specify `-D X10_LIBRARY=` with a full path to the library
and `-D X10_INCLUDE_DIRS=` for the header search path.  This should
allow eliding the x10 build.  The local build can be forced by setting
`USE_BUNDLED_X10` to `NO`.