X10 bindings: appease `cl`
Unfortunately, `clang` accepts the redeclaration with different linkage,
`cl` objects.  This does an even worse workaround for this problem by
relying on the implementation to mark the declarations with
`__declspec(dllexport)` and getting the annotation incorrect for the
user code.  This should have only a minor penalty for Windows in terms
of performance.  Further complicating matters is that `cl` does not seem
to accept `__declspec(dllexport)` after the return type with bazel for
some reason, so explicitly place it before the decl, which would break
clang's `__attribute__` usage.