build: restructure the X10 swift bindings to split X10 and CTensorFlow (#854)
This restructures the "swift_bindings" which is the C++ implementation
of X10 Tensor.  This split allows CTensorFlow and CX10 to be separated
enabling X10 to be used locally when building the X10 Swift interface.

Since CTensorFlow is fully internal, we remove the autolinking it used
previously.  This allows us to use the same module for tensorflow and
x10.