Windows Builds (#863)
* X10: IWYU in X10.TensorWrapper

The TensorWrapper was using types from `stdint.h` without including the
header.  This was caught by `ucrt`.

* X10: correct linkage for XLATensor

A couple of modules were under-specifying their dependencies on `CX10`.
These modules do `@_implementationOnly` imports of the X10 C interfaces.
Ensure that we have marked the dependencies correctly.

* X10: make the C type conversion platform agnostic

The C++ specification does not mandate an underlying type for enums.
Windows uses signed integer types which Unices tend to use unsigned
types.  This requires type-conversion when using the enums in Swift.

* Tests: move the X10 tests under Tests

This movement ensures that `BUILD_TESTING` is properly honoured.  This
is required for Windows.

* tests: rename the tests in CMake

Use a more fitting name for the tests in CMake.  This is purely
aesthetic and has no bearing on the build or testing.