[WIP]: Idea exploring easier layer init.
The problem: when constructing a model, you often need to do some calculations
"out-of-band" / "by hand" to initialize the layers in the model correctly.
For example: in building a simple convolutional network with a dense layer on
top, you need to keep track of how the image is changing through the
convolutions to ensure you set the input size of the dense layer correctly. If
it is set incorrectly, you get an immediate shape mis-match.

The proposed solution: layers are given an extra "shape-based initializer" that
allows the layer's initializer to propagate shape information forward.

Alternatives considered: a number of libraries (e.g. Keras, Haiku) don't
initialize the parameters of libraries until the first run through. This is
cumbersome in Swift, as that would require `func callAsFunction` to be marked
as mutating, or for layers to become `class`es.

Problems of the current design: I just put this together quickly and
specialized everything to `Scalar == Float`. This is obviously suboptimal.

Extensions:
 - *No-op device*: Right now, we perform computations on zeros, just to avoid
   writing duplicated shape-propagation rules. Instead, if we had a no-op
   device, then we could trivially reuse the layer implementations. (Or
   alternatively, if we somehow made `func callAsFunction` generic over the
   Tensor type.)
 - *VMap*: Proper vmap support would simplify these shape-heavy computations.