match precision in gelu  (#965)
This replaces #962

The tensors created in gelu are matched to the precision of the input.

I added X10 tests that would have caught the mismatch. I have not run them since I don't know how to trigger the x10 tests. The added tests are based off leakyRelu X10 tests.

I added a test of gelu gradient to pick up an error that was missed by tests at an intermediate step in these changes.