Docker: delay copying the source tree
This delays the copying of the source tree until the image is setup.  Doing so permits
the previous layers to remain prestine and allows for re-use across multiple builds.
This cache significantly speeds up local testing as the image updates do not need to be
processed.

This comes at the cost of CI validation, which may miss updates to the VM.  This can be
repaired by cleaning the images after building or before building.