build: repair the macOS build (#1050)
Explicitly link against the PythonKit library which is required when
PythonKit is not in the library search path.  Likely we got away with
this due to previous toolchains containing a copy of PythonKit.  This
ensures that we link with the absolute path and use the just built
PythonKit library.

Mark the PythonKit library as being generated as part of the PythonKit
build.  This is required to ensure that CMake understands that the
library is generated and will not exist before the target is built.

Finally, forward the `CMAKE_Swift_COMPILER_TARGET` flag to the
sub-configure for PythonKit.  This is required to ensure that actually
set the target deployment version to the same version as the host.  If
one is not specified, an empty string will be passed which should not
impact the generated compiler invocation.