build: add `INTERFACE_LINK_DIRECTORIES` for `PythonKit` (#1052)
When building on Linux without PythonKit in the toolchain, the
autolinking implementation will add a number of `-lPythonKit` flags to
the linker invocation implicitly.  Because PythonKit is not in the
toolchain and if it is not in the linker search path, the library will
not be resolved by the linker.

This should allow us to find the library irrespective of it being
present in the linker search path or just being built.