Fix test failures caused by `Optional: KeyPathIterable` conformance
https://github.com/apple/swift/pull/33992 added `Optional: KeyPathIterable` conformance.
The added conformance caused code parts like `self[keyPath: kp] as? _KeyPathIterableBase`
to be interpreted as casting `Optional` (return type of `self[keyPath: kp]`) to
`_KeyPathIterableBase` (which always succeeds because of the added conformance).

The original intention was to cast the wrapped value of `Optional` to `_KeyPathIterableBase`.
This original behavior is restored by changing
`if let nested = self[keyPath: kp] as? _KeyPathIterableBase` to
`if let value = self[keyPath: kp], let nested = value as? _KeyPathIterableBase`,
where `value` first "unwraps" `Optional` and `nested` is the result of casting wrapped
value to `_KeyPathIterableBase`.