build: add a hard dependency on PythonKit with SPM (#1151)
Unlike CMake, SPM does not permit any control over build configuration.
Make PythonKit a hard dependency for the package.  This along with the
remainder of the changes that have been made to the build system now
enable building Swift APIs with SPM in a complete configuration of
PythonKit + Swift Numerics + X10.  This simplifies the path towards
getting a SPM built Swift APIs with X10 and PythonKit support with a
standard toolchain.