TensorFlow: silence a warning from the Swift compiler (#1153)
The trailing closure application to the function can be confused with
the function body as per the diagnostic emitted by the swift compiler.
With the recent changes to the application of trailing closures and
support for multiple trailing closures, the syntax is accepted but is
confusing.  Convert from the trailing closure to an inline closure to
make this explicit.