Documentation: update the developer documentation (#1157)
Documentation: update the developer documentation

This adds documentation covering how to build X10 for Windows, Linux,
and macOS as well as links to prebuilt binaries.  Additionally, document
the required additional flags for building Swift APIs.  This is required
in order to build with a toolchain which does not have X10 integrated
into the image.  It also paves the way towards building Swift APIs with
the Swift.org toolchain.

Add a note about SIP and quarantining.

Co-authored-by: Dan Zheng <danielzheng@google.com>
Co-authored-by: Brad Larson <bradlarson@google.com>