documentation: improve explanation for CMake based builds
This adds additional documentation on how to avoid building X10 if the
prebuilt option is being used or if it has been built previously
manually.

Add a note about CMake defaulting to building a copy of X10 if one is
not provided.  Additionally, add examples of what the flags should look
like for various cases to avoid building X10 by default with CMake.