build: attempt to use bash for SWIFTFLAGS (#1173)
Unify the handling of SWIFTFLAGS for stock and non-stock toolchains.
Unfortunately, docker cannot pass along environment variables (outside
of `ARGS`).  We could pass the values through `ARGS` by creating a new
base and switching between the base determined by the toolchain we are
using.  Alternatively, we can create a file to pass along the state,
which I have opted to use here.