build: enable building X10 support with SPM
Enable building the Swift interfaces for X10 with SPM.  This effectively
hard-enables X10 support in TensorFlow as SPM does not provide
configuration.

In order to accomplish this, we need to create symbolic links to sources
in order to merge the `x10` sources into the TensorFlow module.  It is
not possible to simply move the sources to reflect this structure due to
out-of-tree use cases.