build: add a submodule for CX10 handling just the modules
The CX10 module setup is currently very complicated.  The module is
built twice - once via bazel and incorporated into the TensorFlow
library.  The second build is built via Swift for the modules to import
the interfaces into Swift.  In order to build the swift bindings via
SPM, we need to extract the sources.  However, we cannot reorganise the
CX10 subdirectory to enable SPM to filter out just the modules.  As a
result, we create a `CX10Modules` module that is only built when
building with SPM.  CMake builds will ignore this directory, building
the modules from the CX10 directory directly as it can handle the
subdirectory layout.